requirements.txt
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
python-telegram-bot==21.4
requests==2.32.3
setuptools==73.0.1
